--- src/main/java/racingcar/domain/MotorRacingRule.java
@@ -1,5 +1,0 @@
-package racingcar.domain;
-
-public interface MotorRacingRule {
-    boolean pass();
-}
