+++ src/main/java/racingcar/domain/core/NamedRacingCar.java
@@ -2,7 +2,7 @@ package racingcar.domain.core;
 
 import java.util.regex.Pattern;
 
-import racingcar.domain.MotorRacingRule;
+import racingcar.domain.rule.MotorRacingRule;
 import racingcar.domain.RacingResult;
 
 import static racingcar.domain.RacingResult.namedResult;
