+++ src/main/java/racingcar/domain/rule/MotorRacingRandomMoveRule.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.rule;
 
 import java.util.Random;
