+++ src/main/java/racingcar/domain/rule/MotorRacingRule.java
@@ -1,0 +1,8 @@
+package racingcar.domain.rule;
+
+public interface MotorRacingRule {
+    boolean pass();
+    static MotorRacingRule defaultMoveRule(){
+        return new MotorRacingRandomMoveRule();
+    }
+}
