+++ src/test/java/racingcar/domain/core/RacingCarTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import racingcar.domain.MotorRacingRule;
+import racingcar.domain.rule.MotorRacingRule;
 import racingcar.domain.RacingResult;
 
 import static org.assertj.core.api.Assertions.assertThat;
